Extract streaming pipeline into download/pipeline.rs
Move the producer/consumer download pipeline, download pass
execution, and supporting types out of download/mod.rs into a
dedicated pipeline module. Pure code move with visibility
adjustments.

Moved: stream_and_download_from_stream, build_download_outcome,
run_download_pass, download_single_task, AssetDisposition,
ProducerSkipSummary, StreamingResult, PassConfig/PassResult,
flush_pending_state_writes, and ~28 associated tests.

download/mod.rs: 3953 -> 2036 lines, pipeline.rs: 2019 lines.
Overall: mod.rs went from 6258 to 2036 lines across commits 1-4.
